Fix #18 Using container for test
Works to a point, but is using a shell script to build and start the container.  The Gradle docker plugin does not yet support publishing ports, which is a shame because apart from that it worked.  (Code is commented out in build.gradle) .  Closing this issue but opened #19 pending [plugin feature](https://github.com/bmuschko/gradle-docker-plugin/issues/30) becoming available.